app/upload_web/requirements.txt
flask>=3.0,<4
python-dotenv>=1.0,<2
werkzeug>=3.0,<4
